Rename all occurrences of Addressbook to Investigapptor
