+++ src/main/java/seedu/address/commons/events/model/InvestigapptorChangedEvent.java
@@ -1,14 +1,14 @@
 package seedu.address.commons.events.model;
 
 import seedu.address.commons.events.BaseEvent;
-import seedu.address.model.ReadOnlyAddressBook;
+import seedu.address.model.ReadOnlyInvestigapptor;
 
-/** Indicates the AddressBook in the model has changed*/
+/** Indicates the Investigapptor in the model has changed*/
 public class AddressBookChangedEvent extends BaseEvent {
 
-    public final ReadOnlyAddressBook data;
+    public final ReadOnlyInvestigapptor data;
 
-    public AddressBookChangedEvent(ReadOnlyAddressBook data) {
+    public AddressBookChangedEvent(ReadOnlyInvestigapptor data) {
         this.data = data;
     }
