+++ src/test/java/seedu/address/model/crimecase/CaseNameTest.java
@@ -1,0 +1,4 @@
+package seedu.address.model.crimecase;
+
+public class CaseNameTest {
+}
